requirements.txt
streamlit
langchain
openai
pandas
faiss-cpu
tiktoken
python-dotenv
langchain-openai
